refactor: move Type answer functions to a separate file
