+++ AnkiDroid/src/main/java/com/ichi2/anki/previewer/CardViewerViewModel.kt
@@ -40,8 +40,6 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
-import org.jetbrains.annotations.VisibleForTesting
-import org.json.JSONObject
 import timber.log.Timber
 
 abstract class CardViewerViewModel(
@@ -195,54 +193,4 @@ abstract class CardViewerViewModel(
         } else {
             throw IllegalArgumentException("Unhandled POST request: $uri")
         }
-
-    companion object {
-        // ********************************** Type-in answer ************************************
-        /** From the [desktop code](https://github.com/ankitects/anki/blob/1ff55475b93ac43748d513794bcaabd5d7df6d9d/qt/aqt/reviewer.py#L669] */
-        @VisibleForTesting
-        val typeAnsRe = Regex("\\[\\[type:(.+?)]]")
-
-        suspend fun getTypeAnswerField(
-            card: Card,
-            text: String,
-        ): JSONObject? {
-            val match = typeAnsRe.find(text) ?: return null
-
-            val typeAnsFieldName =
-                match.groups[1]!!.value.let {
-                    if (it.startsWith("cloze:")) {
-                        it.split(":")[1]
-                    } else {
-                        it
-                    }
-                }
-
-            val fields = withCol { card.noteType(this).flds }
-            for (i in 0 until fields.length()) {
-                val field = fields.get(i) as JSONObject
-                if (field.getString("name") == typeAnsFieldName) {
-                    return field
-                }
-            }
-            return null
-        }
-
-        suspend fun getExpectedTypeInAnswer(
-            card: Card,
-            field: JSONObject,
-        ): String? {
-            val fieldName = field.getString("name")
-            val expected = withCol { card.note(this@withCol).getItem(fieldName) }
-            return if (fieldName.startsWith("cloze:")) {
-                val clozeIdx = card.ord + 1
-                withCol {
-                    extractClozeForTyping(expected, clozeIdx).takeIf { it.isNotBlank() }
-                }
-            } else {
-                expected
-            }
-        }
-
-        fun getFontSize(field: JSONObject): String = field.getString("size")
-    }
 }
